Updating the Framework
Bot is reading possible greetings out of helloDict.txt and possible answers to his first questions out of howAreYouDict.txt

Currently only limited access, the lists are editable and the bot will always use the newest one once started.

If he doesn't find the answer in the set he will answer for now with "error103"

Code will be commented in the next commit.

So far, bot can read and write.

Need to improve this and add a method to learn from user.